Electrod proposal
