+++ org.alloytools.pardinus.core/src/main/java/kodkod/solvers/api/TemporalSolverFactory.java
@@ -1,0 +1,8 @@
+package kodkod.solvers.api;
+
+import kodkod.engine.TemporalSolver;
+import kodkod.engine.config.ExtendedOptions;
+
+public interface TemporalSolverFactory {
+	TemporalSolver<ExtendedOptions> getTemporalSolver(ExtendedOptions options);
+}
