+++ org.alloytools.pardinus.native/src/main/java/org/alloytools/solvers/natv/electrod/ElectrodSRef.java
@@ -2,12 +2,15 @@ package org.alloytools.solvers.natv.electrod;
 
 import java.util.Optional;
 
+import kodkod.engine.TemporalSolver;
+import kodkod.engine.config.ExtendedOptions;
 import kodkod.engine.satlab.ExternalSolver;
 import kodkod.engine.satlab.SATFactory;
 import kodkod.engine.satlab.SATSolver;
 import kodkod.solvers.api.NativeCode;
+import kodkod.solvers.api.TemporalSolverFactory;
 
-public class ElectrodSRef extends SATFactory {
+public class ElectrodSRef extends SATFactory implements TemporalSolverFactory {
 
 
     private static final long serialVersionUID = 1L;
@@ -60,6 +63,12 @@ public class ElectrodSRef extends SATFactory {
     @Override
     public String type() {
         return "external";
+    }
+
+
+    @Override
+    public TemporalSolver<ExtendedOptions> getTemporalSolver(ExtendedOptions options) {
+        return new ElectrodSolver(options);
     }
